requirements.txt
Flask
flask_cors
flask-security
Flask-SQLAlchemy
mysql-connector-python
marshmallow
uuid
